requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
Django==4.2.2
django-allauth==0.54.0
django-browser-reload==1.11.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-storages==1.13.2
django-extensions
exceptiongroup==1.1.1
google-api-core==2.11.1
google-auth==2.17.3
google-cloud-aiplatform==1.27.1
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.10.0
google-cloud-tasks==2.13.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
model_bakery==1.14.0
nltk==3.8.1
numpy==1.24.4
packaging==23.1
pandas==2.0.2
pdfminer.six==20221105
playwright==1.35.0
pluggy==1.2.0
proto-plus==1.22.3
protobuf==4.23.3
psycopg==3.1.9
psycopg-binary==3.1.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyee==9.0.4
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests_mock==1.11.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
striprtf==0.0.25
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug
whitenoise==6.5.0